+++ examples/src/main/kotlin/io.zenoh/ZBytes.kt
@@ -1,6 +1,8 @@
 package io.zenoh
 
-import io.zenoh.protocol.*
+import io.zenoh.bytes.IntoZBytes
+import io.zenoh.bytes.ZBytes
+import io.zenoh.bytes.into
 import java.nio.ByteBuffer
 import java.nio.ByteOrder
 import kotlin.reflect.typeOf
