+++ zenoh-kotlin/src/commonMain/kotlin/io/zenoh/Session.kt
@@ -20,11 +20,12 @@ import io.zenoh.handlers.ChannelHandler
 import io.zenoh.handlers.Handler
 import io.zenoh.jni.JNISession
 import io.zenoh.keyexpr.KeyExpr
-import io.zenoh.prelude.Encoding
-import io.zenoh.prelude.QoS
-import io.zenoh.protocol.IntoZBytes
-import io.zenoh.protocol.ZBytes
-import io.zenoh.protocol.ZenohId
+import io.zenoh.bytes.Encoding
+import io.zenoh.qos.QoS
+import io.zenoh.bytes.IntoZBytes
+import io.zenoh.bytes.ZBytes
+import io.zenoh.config.Config
+import io.zenoh.config.ZenohId
 import io.zenoh.publication.Delete
 import io.zenoh.publication.Publisher
 import io.zenoh.publication.Put
@@ -33,7 +34,9 @@ import io.zenoh.queryable.Query
 import io.zenoh.queryable.Queryable
 import io.zenoh.sample.Sample
 import io.zenoh.selector.Selector
-import io.zenoh.subscriber.Reliability
+import io.zenoh.qos.Reliability
+import io.zenoh.session.SessionDeclaration
+import io.zenoh.session.SessionInfo
 import io.zenoh.subscriber.Subscriber
 import kotlinx.coroutines.channels.Channel
 import java.time.Duration
