+++ zenoh-kotlin/src/commonMain/kotlin/io/zenoh/bytes/Encoding.kt
@@ -12,7 +12,7 @@
 //   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
 //
 
-package io.zenoh.prelude
+package io.zenoh.bytes
 
 /**
  * Default encoding values used by Zenoh.
