+++ zenoh-kotlin/src/commonMain/kotlin/io/zenoh/bytes/IntoZBytes.kt
@@ -1,4 +1,4 @@
-package io.zenoh.protocol
+package io.zenoh.bytes
 
 /**
  * IntoZBytes interface.
