+++ zenoh-kotlin/src/commonMain/kotlin/io/zenoh/publication/Delete.kt
@@ -15,9 +15,9 @@
 package io.zenoh.publication
 
 import io.zenoh.keyexpr.KeyExpr
-import io.zenoh.prelude.QoS
-import io.zenoh.protocol.ZBytes
-import io.zenoh.subscriber.Reliability
+import io.zenoh.qos.QoS
+import io.zenoh.bytes.ZBytes
+import io.zenoh.qos.Reliability
 
 /**
  * Delete operation.
