+++ zenoh-kotlin/src/commonMain/kotlin/io/zenoh/queryable/Query.kt
@@ -19,11 +19,11 @@ import io.zenoh.selector.Selector
 import io.zenoh.exceptions.ZError
 import io.zenoh.jni.JNIQuery
 import io.zenoh.keyexpr.KeyExpr
-import io.zenoh.prelude.Encoding
-import io.zenoh.prelude.QoS
-import io.zenoh.prelude.SampleKind
-import io.zenoh.protocol.IntoZBytes
-import io.zenoh.protocol.ZBytes
+import io.zenoh.bytes.Encoding
+import io.zenoh.qos.QoS
+import io.zenoh.sample.SampleKind
+import io.zenoh.bytes.IntoZBytes
+import io.zenoh.bytes.ZBytes
 import io.zenoh.sample.Sample
 import org.apache.commons.net.ntp.TimeStamp
