+++ zenoh-kotlin/src/commonMain/kotlin/io/zenoh/sample/SampleKind.kt
@@ -12,7 +12,7 @@
 //   ZettaScale Zenoh Team, <zenoh@zettascale.tech>
 //
 
-package io.zenoh.prelude
+package io.zenoh.sample
 
 /** The kind of sample. */
 enum class SampleKind {
