+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/EncodingTest.kt
@@ -1,9 +1,10 @@
 package io.zenoh
 
 import io.zenoh.keyexpr.intoKeyExpr
-import io.zenoh.prelude.Encoding
-import io.zenoh.protocol.ZBytes
-import io.zenoh.protocol.into
+import io.zenoh.bytes.Encoding
+import io.zenoh.bytes.ZBytes
+import io.zenoh.bytes.into
+import io.zenoh.config.Config
 import io.zenoh.query.ReplyError
 import io.zenoh.sample.Sample
 import io.zenoh.selector.intoSelector
