+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/GetTest.kt
@@ -15,8 +15,9 @@
 package io.zenoh
 
 import io.zenoh.handlers.Handler
-import io.zenoh.prelude.SampleKind
-import io.zenoh.protocol.into
+import io.zenoh.sample.SampleKind
+import io.zenoh.bytes.into
+import io.zenoh.config.Config
 import io.zenoh.query.Reply
 import io.zenoh.queryable.Queryable
 import io.zenoh.selector.Parameters
