+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/PublisherTest.kt
@@ -15,10 +15,11 @@
 package io.zenoh
 
 import io.zenoh.keyexpr.KeyExpr
-import io.zenoh.prelude.Encoding
+import io.zenoh.bytes.Encoding
 import io.zenoh.keyexpr.intoKeyExpr
-import io.zenoh.prelude.SampleKind
-import io.zenoh.protocol.into
+import io.zenoh.sample.SampleKind
+import io.zenoh.bytes.into
+import io.zenoh.config.Config
 import io.zenoh.publication.Publisher
 import io.zenoh.sample.Sample
 import io.zenoh.subscriber.Subscriber
