+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/QueryableTest.kt
@@ -14,15 +14,20 @@
 
 package io.zenoh
 
+import io.zenoh.bytes.Encoding
 import io.zenoh.handlers.Handler
 import io.zenoh.keyexpr.KeyExpr
 import io.zenoh.keyexpr.intoKeyExpr
-import io.zenoh.prelude.*
-import io.zenoh.protocol.into
+import io.zenoh.bytes.into
+import io.zenoh.config.Config
+import io.zenoh.qos.CongestionControl
+import io.zenoh.qos.Priority
+import io.zenoh.qos.QoS
 import io.zenoh.query.Reply
 import io.zenoh.query.ReplyError
 import io.zenoh.queryable.Query
 import io.zenoh.sample.Sample
+import io.zenoh.sample.SampleKind
 import io.zenoh.selector.Selector
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.channels.Channel
