+++ zenoh-kotlin/src/commonTest/kotlin/io/zenoh/SubscriberTest.kt
@@ -17,12 +17,13 @@ package io.zenoh
 import io.zenoh.handlers.Handler
 import io.zenoh.keyexpr.KeyExpr
 import io.zenoh.keyexpr.intoKeyExpr
-import io.zenoh.prelude.Encoding
+import io.zenoh.bytes.Encoding
 import io.zenoh.sample.Sample
-import io.zenoh.prelude.CongestionControl
-import io.zenoh.prelude.Priority
-import io.zenoh.prelude.QoS
-import io.zenoh.protocol.into
+import io.zenoh.qos.CongestionControl
+import io.zenoh.qos.Priority
+import io.zenoh.qos.QoS
+import io.zenoh.bytes.into
+import io.zenoh.config.Config
 import kotlinx.coroutines.DelicateCoroutinesApi
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.runBlocking
